Adds prefix for tracking changes to documentation
This is meant to track changes to files such as `README*` or anything that might appear in a `docs` folder. The intent is to use this prefix for documentation changes which have absolutely no effect on the output of the code. It should not be used for making changes to any documentation that is rendered to a user of code in question. For example changes to the output produced by running `--help` for a console application *should not* use this prefix. An upper-case prefix should be used for such changes.